Fix: Changed all `subscript(…)` properties' `index` arg back to `Int`-type
Because of find-and-replace, the `index` arg got changed to `Int32` for some types and `Float` for others.

* Changed all `subscript` properties back to the standard `subscript(index:Int) -> …` signature.

+++ Sources/Float2/Float2.swift
@@ -99,7 +99,7 @@ extension Float2
 	// MARK: `subscript`-Getter
 	
 	/// Access individual elements of the vector via subscript.
-	public subscript(index:Float) -> Float {
+	public subscript(index:Int) -> Float {
 		switch index {
 			case 0: return self.x
 			case 1: return self.y
